added spliced assets
This took way longer than it should have. The main issue was the fact that the Scott Pilgrim spritesheets are non-uniform and therefore cannot be cut by any grid-based slicer. So, obviously, I perused the internet, found nothing, then wrote my own script to slice a non-uniform spritesheet. It's included in the Robot_01 sliced folder.