setup controls + input wishlist
I want input to be a priority-queue thing. Unfortunately, Unity doesn't have a recent version of .NET, so I had to port the source file into the project (fixing the various compiler errors).

My general idea for input is to have it be essentially just a buffered list of input histories with "priority" inputs beating out other ones. A modular way to do this would be to have SOs arranging a list of input types into order to establish priorities. I can then use these priorities when adding elements to the PriorityQueue, since the queue separates the "value" from the "priority".

Currently, InputBufferReader is a tentative class. Right now, it was used to debug PriorityQueue.