Fixed state quad-fire bug
The issue (aside from my bad frame math) was that state machine behaviours are cloned upon creation, resulting in inconsistencies. OnStateEnter, for example, was called multiple times. To fix this, I had to add the attribute flag to prevent clones of behaviours. I also had to add a boolean flag to prevent multiple entering. Another minor bugfix was the correct of CeilToInt to FloorToInt, as that prevented the scrubbing from being one frame ahead.

I created some sample tracks which can be seen in-game. Currently they operate with exit time (which won't be the case in the future), but the demonstration works. The Tracks even support keys with the same frame values, meaning they all get executed at once. You can also add a SequenceBehavior to every state w/o issue; I thought that the attribute would prevent that, but it still works.

Next steps are to actually implement the listeners for Tracks, the Track types themselves (i.e. flag track, translation track, hitbox track, etc.), the player controllers, and basically the rest of the game (lol).