added TextMeshPro, partial hitbox, visualizers
Added a partial implementation of hitboxes, a hitbox track, and a training hitbox visualizer. Reworked the TrackSO to use an enum instead of raw types when requesting component listeners. Additionally, abstracted ComponentMaps for future polymorphism and so I could make a lightweight training component map.

Training/Visualization mode is now operational! I need to add more tracks besides Hitbox, but that can come later. There were a lot of little bugs involving animator playback, but the main one was animation being off by a frame. The solution was pretty simple (albeit heavy-handed), but you just had to set the playback time to a bit beyond the keyframe, so that it would get evaluated.

The MonoTrainingManager needs to be abstracted, as it is currently doing a lot. Problem is, all the things it's doing are so small that it might just make it harder to read to have them be in different Monos. I could try adding some raw classes, though.