better implementation of looping
Sleep really does fix everything.

Implementation has been reworked to use conditionals. I got around the issues I mentioned in the previous commit in one major (albeit heavy-handed) way; PredicateSOs wholly store conditional logic. All the listeners have to do is get the conditional, pass it a component map, and let it return a value. This goes against how I've been handling the data-blocks in Tracks in the sense that the "computing" part of the track is not the listener but rather the Predicate on the data-blocks. However, this minor violation serves a better structure than what I had before. Now, the listeners function as Predicate-firing behaviors that rewind the Sequence if the predicate is true.
I avoided the polymorphism issue by hiding the "generic" behavior inside of a protected abstract method (Eval) that extending Predicates would implement. That way, they would have access to a ComponentMap (through the genericless component map retriever) to allow them to perform operations on different component types but still have the required method Check() to actually get a usable boolean result. Wrapping my head around the encapsulation shenanigans was a bit strange, but it works out.

This allowed me to free the coupling between the conditional evaluator and monoinputcontroller. However, the side-effect of this is that ComponentMap has drifted from its initial purpose, slightly. Instead of being a component resource for Tracks and Sequences to setup their listeners, it has become more of a general "here is a place for components" class. Currently, only PredicateSO is forcing this new definition which is good. I'd rather it not scale any harder than this because the ComponentType enum would get very cluttered, since now it has "typedefs" for non-evaluator things (MonoInputController). Risky.

All in all, way better than what I had yesterday. Good grief and alas.